% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-1.R
\name{des1_3_1}
\alias{des1_3_1}
\title{algorithm 1.3.1: compute discrete time evolution of inventory level for simple system}
\usage{
des1_3_1(demands, s, S)
}
\arguments{
\item{demands}{vector of demands}

\item{s}{minimum inventory level}

\item{S}{maximum inventory level}
}
\value{
a list with \code{l} (inventory levels) and \code{o} (orders)
}
\description{
If the demands d1, d2, . . . are known then this algorithm computes the discrete time evolution of the inventory level for a simple (s, S) inventory system with back ordering and no delivery lag.
}
\examples{
d <- c(30,15,25,15,45,30,25,15,20,35,20,30)
s <- 20
S <- 60
des1_3_1(d,s,S)
}
