% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-1.R
\name{des_1_2_1}
\alias{des_1_2_1}
\title{algorithm 1.2.1: calculate delays under FIFO with finite capacity}
\usage{
des_1_2_1(a, s)
}
\arguments{
\item{a}{vector of arrival times}

\item{s}{vector of service times}
}
\value{
d vector of delay times (how long each job waits in the queue prior to begin of service)
}
\description{
If the arrival times a1, a2, . . . and service times s1, s2, . . . are known and if the server is initially idle, then this algorithm computes the delays d1,d2,... in a single-server FIFO service node with infinite capacity.
This algorithm computes the departure times c1, c2, . . . as a by-product of the computation.
}
\examples{
a <- c(15,47,71,111,123,152,166,226,310,320)
s <- c(43,36,34,30,38,40,31,29,36,30)
des_1_2_1(a,s)
}
