% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-2.R
\name{des_2_1_1}
\alias{des_2_1_1}
\title{algorithm 2.1.1: determine if \code{a} is full period multiplier relative to prime modulus \code{m}}
\usage{
des_2_1_1(a, m)
}
\arguments{
\item{a}{multiplier, a fixed integer}

\item{m}{modulus, a fixed large prime integer}
}
\value{
\code{true} or \code{false} if \code{a} is a full period multiplier
}
\description{
This algorithm, based upon Theorem 2.1.2, can be used to determine
if a is a full-period multiplier relative to the prime modulus m.
If m is not prime Algorithm 2.1.1 may not halt, although the while loop
will occasionally check for user interrupts.
}
\examples{
des_2_1_1(7,13) # true
des_2_1_1(5,13) # false
}
