% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-2.R
\name{des_2_2_1}
\alias{des_2_2_1}
\title{algorithm 2.2.1: evaluate ax mod m without producing any integers larger than m-1}
\usage{
des_2_2_1(x, a, m)
}
\arguments{
\item{x}{value in \{1,2,...,m-1\}}

\item{a}{multiplier, a fixed integer}

\item{m}{modulus, a fixed large prime integer}
}
\value{
value of function
}
\description{
If m = aq+r is prime, r < q, and x ∈ Xm, then g(x) = ax mod m can be evaluated as follows without producing any intermediate or final values larger than m − 1 in magnitude.
}
\examples{
 # gives same result
 des_2_2_1(x = 9,a = 3,m = 401)
 (3*9) \%\% 401
 # to avoid overflow
m <- as.integer((2^31)-1)
a <- 48271
x <- 53423L
(a*x) \%\% m
des_2_2_1(x,a,m)
}
