% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-2.R
\name{des_2_5_1}
\alias{des_2_5_1}
\title{algorithm 2.5.1 Given the state transition function g(·) and initial state x0, this algorithm determines the fundamental pair (s, p)}
\usage{
des_2_5_1(g, x0)
}
\arguments{
\item{g}{function}

\item{x0}{initial state}
}
\value{
a vector of all other full-period modulous-compatible multipliers
}
\description{
algorithm 2.5.1 Given the state transition function g(·) and initial state x0, this algorithm determines the fundamental pair (s, p)
}
\examples{
midsq <- midsq <- function(x){as.integer(floor((x^2)/100) \%\% 10000)}
x0 <- 1726
des_2_5_1(midsq,x0)
}
