% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-2.R
\name{des_2_5_2}
\alias{des_2_5_2}
\title{algorithm 2.5.2 Given the state transition function g(·) and initial state x0, this algorithm determines the fundamental pair (s, p)}
\usage{
des_2_5_2(g, x0)
}
\arguments{
\item{g}{function}

\item{x0}{initial state}
}
\value{
fundamental pair (starting point, period)
}
\description{
This algorithm trades slow speed for low memory; all intermediate
values are generated on the fly via calls to \code{g}.
}
\examples{
midsq6 <- function(x){as.integer(floor((x^2)/1000) \%\% 1000000)}
x0 <- 141138 # this x0 will give (s,p) = (296, 29)
x0 <- 119448 # this x0 will give (s,p) = (428, 210)
x0 <- 586593 # this x0 will give (s,p) = (48, 13)
x0 <- 735812 # this x0 will give (s,p) = (225, 1)
x0 <- 613282 # this x0 will give (s,p) = (469, 20)
des_2_5_2(midsq6,x0)
}
