% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-4.R
\name{des_4_1_1}
\alias{des_4_1_1}
\title{algorithm 4.1.1: Welford's one pass algorithm}
\usage{
des_4_1_1(sample)
}
\arguments{
\item{sample}{a vector of values (will be coerced by \code{\link{as.numeric}})}
}
\value{
a vector with sample mean and standard deviation
}
\description{
Calculate the sample mean and standard deviation in one pass via
Welford's algorithm.
}
\examples{
samp <- rnorm(n=1e4)
mean(samp);sd(samp)
des_4_1_1(samp)
}
