% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-4.R
\name{des_4_2_1}
\alias{des_4_2_1}
\title{algorithm 4.2.1: discrete data histrogram}
\usage{
des_4_2_1(a, b, data)
}
\arguments{
\item{a}{lower bound}

\item{b}{upper bound}

\item{data}{a vector of values (will be coerced by \code{\link{as.integer}})}
}
\value{
a list with \code{count}, \code{outliers_lo}, and \code{outliers_hi}
}
\description{
Given integers a, b with a < b and integer-valued data x1, x2, . . . the following algorithm computes a discrete-data histogram.
}
\examples{
samp <- rpois(n=30,lambda=10)
des_4_2_1(a=2,b=35,data=samp)
}
