% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-errata.R
\name{des_gcd}
\alias{des_gcd}
\title{Greatest Common Divisor (GCD)}
\usage{
des_gcd(a, b)
}
\arguments{
\item{a}{a positive integer}

\item{b}{a positive integer}
}
\value{
the GCD
}
\description{
Euclidean Algorithm to compute GCD.
}
\examples{
des_gcd(10,21) # returns 1
des_gcd(12,24) # returns 12
}
