% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-errata.R
\name{des_sieve}
\alias{des_sieve}
\title{Sieve of Eratosthenes}
\usage{
des_sieve(N)
}
\arguments{
\item{N}{a positive integer greater than 2}
}
\value{
a vector of prime numbers
}
\description{
Finds all prime numbers between 2 and N (N>2)
}
\examples{
# returns 25 prime numbers
des_sieve(100)
}
