% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-1.R
\name{des_sis1}
\alias{des_sis1}
\title{program sis1: simulates a simple (s,S) inventory system using trace-driven demand}
\usage{
des_sis1(demands, s, S)
}
\arguments{
\item{demands}{vector of demands}

\item{s}{minimum inventory level}

\item{S}{maximum inventory level}
}
\value{
a named vector with average demand, order, setup frequency, holding level, and shortage level
}
\description{
This program simulates a simple (s,S) inventory system using demand read
from vector input.  Backlogging is permitted and there is no delivery lag.
The output statistics are the average demand and order per time interval
(they should be equal), the relative frequency of setup and the time
averaged held (+) and short (-) inventory levels.
}
\examples{
data(sis1dat)
round(des_sis1(sis1dat$d,20,80),3)
}
