% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des-2.R
\name{des_2_5_1}
\alias{des_2_5_1}
\title{algorithm 2.5.1 Given the state transition function g(·) and initial state x0, this algorithm determines the fundamental pair (s, p)}
\usage{
des_2_5_1(g, x0)
}
\arguments{
\item{g}{function}

\item{x0}{initial state}
}
\value{
fundamental pair (starting point, period)
}
\description{
This algorithm trades memory requirements for fast speed; all intermediate
values are stored in a linked list to minimize calls to \code{g}.
}
\examples{
midsq <- function(x){as.integer(floor((x^2)/100) \%\% 10000)}
x0 <- 1726 # this x0 will give (s,p) = (6,4)
x0 <- 6283 # this x0 will give (s,p) = (7,1)
x0 <- 5600 # this x0 will give (s,p) = (0,4)
des_2_5_1(midsq,x0)
}
